update deps and rearrange modules
