+++ api/Cargo.toml
@@ -7,18 +7,19 @@ authors = ["@anorth"]
 edition = "2021"
 
 [dependencies]
-fvm_shared = { version = "3.2.0", default-features = false }
-fvm_ipld_hamt = "0.6.1"
-fvm_ipld_amt = { version = "0.5.1", features = ["go-interop"] }
-fvm_ipld_car = { version = "~0.6.0"}
-fvm_ipld_blockstore = { version = "~0.1.1"}
-fvm_ipld_encoding = "0.3.3"
-fil_actors_runtime = { version = "12.0.0", git = "https://github.com/filecoin-project/builtin-actors", branch = "master", features = ["test_utils"] }
+fvm_shared = { version = "3.4.0", default-features = false }
+fvm_ipld_hamt = "0.7.0"
+fvm_ipld_amt = { version = "0.6.0", features = ["go-interop"] }
+fvm_ipld_car = { version = "~0.7.0"}
+fvm_ipld_blockstore = { version = "0.2.0"}
+fvm_ipld_encoding = "0.4.0"
 
 anyhow = "~1.0.47"
 blake2b_simd = { version = "1.0" }
 bimap = "0.6.2"
-cid = { version = "~0.8.5", default-features = false }
+cid = { version = "0.10.1", default-features = false }
 itertools = "~0.10.5"
+libsecp256k1 = { version = "0.7.1"}
+multihash = { version = "0.18.1", default-features = false, features=["sha2", "derive"] }
 num-format = "0.4.4"
 thiserror = "~1.0.30"
