+++ builtin/Cargo.toml
@@ -28,22 +28,22 @@ fil_actors_runtime = { version = "12.0.0", git = "https://github.com/filecoin-pr
 
 
 fvm = { version = "3.2.0", default-features = false, features = ["testing"] }
-fvm_actor_utils = { version = "6.0.0" } 
-fvm_shared = { version = "3.2.0", default-features = false }
-fvm_ipld_hamt = "0.6.1"
-fvm_ipld_amt = { version = "0.5.1", features = ["go-interop"] }
-fvm_ipld_car = { version = "~0.6.0"}
-fvm_ipld_blockstore = { version = "~0.1.1"}
-fvm_ipld_encoding = "0.3.3"
+fvm_actor_utils = { version = "7.0.0" } 
+fvm_shared = { version = "3.4.0", default-features = false }
+fvm_ipld_hamt = "0.7.0"
+fvm_ipld_amt = { version = "0.6.0", features = ["go-interop"] }
+fvm_ipld_car = { version = "~0.7.0"}
+fvm_ipld_blockstore = { version = "0.2.0"}
+fvm_ipld_encoding = "0.4.0"
 fvm_ipld_bitfield = { version = "~0.5.2"}
 
 # Used directly in some tests
-frc46_token = "6.0.0"
+frc46_token = "7.0.0"
 
 anyhow = "~1.0.47"
 blake2b_simd = "1.0.0"
 bls-signatures = { version = "0.13", default-features = false }
-cid = { version = "~0.8.5", default-features = false }
+cid = { version = "~0.10.1", default-features = false }
 libsecp256k1 = { version = "0.7" }
 multihash = { version = "~0.16.1", default-features = false }
 rand_chacha = "~0.3"
