+++ builtin/tests/util/deals.rs
@@ -4,7 +4,6 @@ use fil_actor_market::{
 };
 use fil_actors_runtime::cbor::serialize;
 use fil_actors_runtime::runtime::Policy;
-use fil_actors_runtime::test_utils::make_piece_cid;
 use fil_actors_runtime::STORAGE_MARKET_ACTOR_ADDR;
 use fvm_shared::address::Address;
 use fvm_shared::clock::ChainEpoch;
@@ -12,6 +11,7 @@ use fvm_shared::crypto::signature::{Signature, SignatureType};
 use fvm_shared::econ::TokenAmount;
 use fvm_shared::error::ExitCode;
 use fvm_shared::piece::PaddedPieceSize;
+use fvm_workbench_api::vm::test_utils::make_piece_cid;
 use fvm_workbench_api::wrangler::VM;
 use num_traits::Zero;
