+++ vm/Cargo.toml
@@ -11,16 +11,16 @@ fvm-workbench-api = { path = "../api" }
 
 # fvm feature=testing required to not be amazingly slow
 fvm = { version = "3.2.0", default-features = false, features = ["testing"] }
-fvm_shared = { version = "3.2.0", default-features = false }
-fvm_ipld_hamt = "0.6.1"
+fvm_shared = { version = "3.4.0", default-features = false }
+fvm_ipld_hamt = "0.7.0"
 fvm_ipld_amt = { version = "0.5.1", features = ["go-interop"] }
-fvm_ipld_car = { version = "~0.6.0"}
-fvm_ipld_blockstore = { version = "~0.1.1"}
-fvm_ipld_encoding = "0.3.3"
+fvm_ipld_car = { version = "~0.7.0"}
+fvm_ipld_blockstore = { version = "0.2.0"}
+fvm_ipld_encoding = "0.4.0"
 replace_with = "0.1.7"
 
 anyhow = "~1.0.47"
-cid = { version = "~0.8.5", default-features = false }
+cid = { version = "0.10.1", default-features = false }
 futures = "~0.3.19"
-multihash = { version = "~0.16.1", default-features = false }
+multihash = { version = "0.18.1", default-features = false }
 thiserror = "~1.0.30"
